core/generator.py
import gspread
import requests
import re
from datetime import datetime
from google.oauth2.service_account import Credentials
import google.auth.transport.requests

SERVICE_ACCOUNT_FILE = 'service_account.json'
TARGET_MODEL = 'models/gemini-2.5-flash'
SHEET_NAME = "Shedevrum_Trends"

def get_gemini_response(text_prompt):
    scopes = ['https://www.googleapis.com/auth/generative-language']
    creds = Credentials.from_service_account_file(SERVICE_ACCOUNT_FILE, scopes=scopes)
    auth_req = google.auth.transport.requests.Request()
    creds.refresh(auth_req)
    url = f"https://generativelanguage.googleapis.com/v1/{TARGET_MODEL}:generateContent"
    headers = {"Authorization": f"Bearer {creds.token}", "Content-Type": "application/json"}
    payload = {"contents": [{"parts": [{"text": text_prompt}]}]}
    response = requests.post(url, headers=headers, json=payload)
    if response.status_code == 200:
        return response.json()['candidates'][0]['content']['parts'][0]['text']
    return "Ошибка API"

def clean_output(text):
    # Убираем все, что похоже на вводные фразы или анализ
    text = text.replace("**", "").replace("__", "")
    if ":" in text[:60]: text = text.split(":", 1)[1].strip()
    
    # Дополнительная жесткая чистка вводных фраз
    phrases = ["вот промпт", "вот вариант", "держи промпт", "конечно", "готово", "промпт для шедеврума", "промпт"]
    text_lower = text.lower()
    for p in phrases:
        if text_lower.startswith(p):
            text = text[len(p):].strip()
            text = re.sub(r'^[,.\-!:\s]+', '', text)
            text_lower = text.lower()
            
    return text.strip()[:480]

try:
    scope = ["https://www.googleapis.com/auth/spreadsheets", "https://www.googleapis.com/auth/drive"]
    creds = Credentials.from_service_account_file(SERVICE_ACCOUNT_FILE, scopes=scope)
    gc = gspread.authorize(creds)
    sheet = gc.open(SHEET_NAME).get_worksheet(0)
    
    all_values = sheet.get_all_values()
    headers = all_values[0]
    rows = all_values[1:]
    col = {name.strip(): i for i, name in enumerate(headers)}

    # Обработка всех строк со статусом New или Redo_Requested
    for i, row in enumerate(rows):
        actual_row = i + 2
        row_padded = row + [''] * (len(headers) - len(row))
        status = row_padded[col["Status"]].strip() if "Status" in col else ""
        prompt_ai_val = row_padded[col["Prompt_AI"]].strip() if "Prompt_AI" in col else ""
        
        # Условие обработки: статус New или Redo_Requested (или пустой Prompt_AI для обратной совместимости)
        if "New" in status or "Redo" in status or (not prompt_ai_val and "Generated" not in status and "Posted" not in status):
            orig_prompt = row_padded[col["Prompt"]] if "Prompt" in col else ""
            orig_model = row_padded[col["Model"]] if "Model" in col else "v2.5"
            aspect_ratio = row_padded[col["Aspect_Ratio"]] if "Aspect_Ratio" in col else "1:1"
            error_log = row_padded[col["Error_Log"]] if "Error_Log" in col else ""
            
            print(f"📈 Обработка строки {actual_row}: {orig_prompt[:30]}... [Модель: {orig_model}] [Формат: {aspect_ratio}]")

            extra_redo_instr = ""
            if "Redo" in status and error_log:
                extra_redo_instr = f"\nВНИМАНИЕ: Этот промпт был отклонен цензурой Шедеврума. Причина: {error_log}. Перепиши его, сделав максимально безопасным, убери любые потенциально стоп-слова, сохранив при этом художественный стиль и детализацию."

            instr = f"""Действуй как эксперт по обходу фильтров Шедеврума. 
Твоя цель — улучшить оригинал: '{orig_prompt}', сохранив 100% его смысла и стиля, но заменив все потенциально 'опасные' для модерации слова на технические и архитектурные термины.
{extra_redo_instr}
ЗАПРЕТ: слияние, переливы, тело, сочный, нежный, плоть, влажный, проникновение.
ИСПОЛЬЗУЙ: оптические эффекты, геометрию, профессиональное освещение (Ray Tracing, Global Illumination, Subsurface Scattering, Chromatic Aberration).
ФОРМАТ ИЗОБРАЖЕНИЯ: {aspect_ratio}.
ВЫХОД: Только текст промпта без пояснений, до 450 знаков, на русском. Завершай полной точкой."""

            ai_text = clean_output(get_gemini_response(instr))
            current_date = datetime.now().strftime("%d.%m.%Y %H:%M")
            
            # Обновляем Prompt_AI, Model_AI, Author_AI ... Status (J-R)
            update_j_r = [[ai_text, orig_model, "Serik AI", "", "", "", "", current_date, "🆕 Generated"]]
            sheet.update(values=update_j_r, range_name=f"J{actual_row}:R{actual_row}")
            
            # Очищаем лог ошибок в колонке U
            if "Error_Log" in col:
                sheet.update(values=[[""]], range_name=f"U{actual_row}")
                
            print(f"✅ Готово для строки {actual_row}")
            
except Exception as e:
    print(f"❌ Системная ошибка: {e}")