requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
playwright
gspread
google-auth
requests
pydantic
pydantic-settings
